+++ src/main/java/lottogame/domain/TicketMachine.java
@@ -3,6 +3,7 @@ package lottogame.domain;
 import lottogame.domain.enums.LottoGameRank;
 import lottogame.domain.lotto.LottoResult;
 import lottogame.domain.lotto.LottoTicket;
+import lottogame.domain.strategy.LottoNumberGenerationStrategy;
 
 public class TicketMachine {
     public LottoTicket createLottoTicket(LottoNumberGenerationStrategy strategy) {
