+++ src/main/java/lottogame/domain/strategy/RangeLottoNumberPickerStrategy.java
@@ -1,4 +1,4 @@
-package lottogame.domain;
+package lottogame.domain.strategy;
 
 import static java.util.stream.Collectors.*;
